+++ sdk/core/azure-core/src/io/body_stream.cpp
@@ -19,6 +19,7 @@
 #endif
 
 #include "azure/core/context.hpp"
+#include "azure/core/internal/io/null_body_stream.hpp"
 #include "azure/core/io/body_stream.hpp"
 
 #include <algorithm>
@@ -218,3 +219,11 @@ size_t ProgressBodyStream::OnRead(
 }
 
 int64_t ProgressBodyStream::Length() const { return m_bodyStream->Length(); }
+
+using Azure::Core::IO::_internal::NullBodyStream;
+
+NullBodyStream* NullBodyStream::GetNullBodyStream()
+{
+  static NullBodyStream nullBodyStream;
+  return &nullBodyStream;
+}