+++ sdk/keyvault/azure-security-keyvault-certificates/test/ut/certificate_client_base_test.hpp
@@ -276,7 +276,13 @@ namespace Azure {
       EXPECT_EQ(
           result.Value.Policy.LifetimeActions[0].LifetimePercentage.Value(),
           action.LifetimePercentage.Value());
-
+      EXPECT_EQ(result.Value.Policy.KeyUsage.size(), size_t(2));
+      auto keyUsage = result.Value.Policy.KeyUsage;
+      EXPECT_TRUE(
+          (keyUsage[0] == CertificateKeyUsage::DigitalSignature
+           && keyUsage[1] == CertificateKeyUsage::KeyEncipherment)
+          || (keyUsage[1] == CertificateKeyUsage::DigitalSignature
+              && keyUsage[0] == CertificateKeyUsage::KeyEncipherment));
       return result.Value;
     }