+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/mocked_transport_adapter_test.hpp
@@ -24,7 +24,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys { nam
 
   namespace _detail {
     // Return a simple key as response so keyvault can parse it to create the T response
-    // Fake key from https://docs.microsoft.com/en-us/rest/api/keyvault/GetKey/GetKey#examples
+    // Fake key from https://docs.microsoft.com/rest/api/keyvault/GetKey/GetKey#examples
     static const char FakeKey[]
         = "{  \"key\": {    \"kid\": "
           "\"https://myvault.vault.azure.net/keys/CreateSoftKeyTest/"