+++ sdk/storage/azure-storage-queues/inc/azure/storage/queues/queue_options.hpp
@@ -66,7 +66,7 @@ namespace Azure { namespace Storage { namespace Queues {
      * operation against another host. If SecondaryHostForRetryReads is "" (the default) then
      * operations are not retried against another host. NOTE: Before setting this field, make sure
      * you understand the issues around reading stale & potentially-inconsistent data at this
-     * webpage: https://docs.microsoft.com/en-us/azure/storage/common/geo-redundant-design.
+     * webpage: https://docs.microsoft.com/azure/storage/common/geo-redundant-design.
      */
     std::string SecondaryHostForRetryReads;