+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 1.4.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.3.1 (2021-11-05)
+
+### Bugs Fixed
+
+- Fixed linking error when Azure SDK is being built as DLL.
+
 ## 1.3.0 (2021-11-04)
 
 ### Features Added