+++ sdk/keyvault/azure-security-keyvault-certificates/CHANGELOG.md
@@ -1,7 +1,20 @@
 # Release History
 
-## 4.0.0-beta.1 (Unreleased)
+## 4.0.0-beta.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 4.0.0-beta.1 (2021-11-09)
 
 ### New Features
 
-- Added `CertificateClient` with  `GetCertificate()` and `GetCertificateVersion()` APIs.
+- Initial beta release of Azure Security Key Vault Certificates API for CPP.
+  - Added `Azure::Security::KeyVault::Certificates::CertificateClient` for get, create, list, delete, backup, restore, and import certificate operations.
+  - Added high-level and simplified `keyvault_certificates.hpp` file for simpler include experience for customers.
+  - Added model types which are returned from the `CertificateClient` operations, such as `Azure::Security::KeyVault::Certificates::KeyVaultCertificate`.