+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -1,12 +1,18 @@
 # Release History
 
-## 12.3.0-beta.1 (Unreleased)
+## 12.3.0 (Unreleased)
 
 ### Features Added
+
+- New API: `BlobClient::CopyFromUri()`.
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 12.2.1 (2021-11-08)
 
 ### Other Changes